LF-5171 Extract the visual parts of FilterMultiSelectV2 into their own component
Now FilterMultiSelectV2 is a filter state wrapper around CheckboxMultiSelect
